asdf

+++ src/main/java/kr/ac/kopo/jeonse/domain/jeonse/domain/AppropriateJeonse.java
@@ -21,7 +21,7 @@ public class AppropriateJeonse {
         private int school;
         private int publicSecurity;
         private int busStop;
-        private int subway;
+        private Float subway;
         private int mart;
     }
 }
